Feature/create account service

See merge request ctd/brasil/projeto-integrador-2/0223/back-end/grupo-1!4

+++ iam-service/src/main/java/com/digitalmoneyhouse/iamservice/service/JwtTokenService.java
@@ -1,5 +1,6 @@
 package com.digitalmoneyhouse.iamservice.service;
 
+import com.digitalmoneyhouse.iamservice.exception.InvalidTokenException;
 import com.digitalmoneyhouse.iamservice.model.JwtToken;
 import com.digitalmoneyhouse.iamservice.repository.JwtTokenRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,8 +22,11 @@ public class JwtTokenService {
         return true;
     }
 
-    public void delete(String token){
+    public void delete(String token) throws InvalidTokenException {
         JwtToken jwtToken = new JwtToken(token);
+        if (!isValid(token)) {
+            throw new InvalidTokenException();
+        };
         repository.delete(jwtToken);
     }
 }
